Conflicts:
	website/static/website/js/ajax.js
	website/templates/order.html

+++ website/static/website/js/ajax.js
@@ -20,30 +20,37 @@ $(document).ready(function() {
 	$("#makeOrder").click(function () {
 
 		cart = sessionStorage.getItem("cart");
-		
+
 		var table_num=parseInt($("#tableNumber").val());
 		var table_num_span = parseInt($("#tableNumber").text());
 		if (isNaN(table_num) && isNaN(table_num_span)){
 			alert("You have to choice table!")
 			return false;
 		}
-		$.ajax({
-			type:"POST",
-			url:"/order/makecurrentorder/",
-			data: {
-				'cart': cart,
-				'table': table_num
-			},
-			success: function(result){
-				if(result.success){
-					sessionStorage.clear();
-					window.location = 'finalize';
-				}
-			},
-			headers: {
-				'X-CSRFToken': csrftoken
-			}
-		})
+
+        if (cart != null){
+
+    		$.ajax({
+    			type:"POST",
+    			url:"/order/makecurrentorder/",
+    			data: {
+    				'cart': cart,
+    				'table': table_num
+    			},
+    			success: function(result){
+    				if(result.success){
+    					sessionStorage.clear();
+                        window.location = 'finalize';
+    				}
+    			},
+    			headers: {
+    				'X-CSRFToken': csrftoken
+    			}
+    		})
+        }
+        else {
+            window.location = 'finalize';
+        }
 	});
